RepoMains/requirements.txt
numpy
pandas
scikit-learn
matplotlib
pyyaml
requests
torch
tensorflow
seaborn
tabulate
